% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_data.R
\name{agg_data}
\alias{agg_data}
\title{agg_data}
\usage{
agg_data(data, thre_tem, thre_visul)
}
\arguments{
\item{data}{The data to be passed onto agg_data function}

\item{thre_tem}{The upper threshold of temperature to decide extreme temperature conditions. The unit of thre_tem is 1 cesius degree. The default value is 0 cesius degree.}

\item{thre_visul}{The upper threshold of visibility distance to decide extreme temperature conditions. The unit of thre_visul is 1 meter. The default value is 1000 meters.}
}
\description{
This function provides a way to aggregate the data onto the hour level.
}
