% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_xls.R
\name{read_all_excel}
\alias{read_all_excel}
\title{read_all_xls}
\usage{
read_all_excel(path = getwd(), add_variable = TRUE)
}
\arguments{
\item{path}{The path from which you want to read all the Microsoft Excel files (with the extension of .xlsx OR .xls). The default value of the parameter path is the current working directory.}

\item{add_variable}{If the add_variable is TRUE, then an extra variable "file_name" will be added to each data.frame in the output list. The default value is set as TRUE.}
}
\value{
a list of dataframes, which can be combined using the rbindlist function in data.table package.
}
\description{
This function provides a way of reading all Microsoft Excel files (.xlsx & .xls) in the folder.
}
\examples{
# Let's assume there are three files (A.xlsx, B.xlsx, C.xls) in the path "C:/data".
all_excel = read_all_excel(path = "C:/data")
# The above code should read all .xlsx & .xls files into the list "all_excel". You can access each data.frame by all_excel[["A"]], all_excel[["B"]], and all_excel[["C"]].
# If the parameter "add_variable" is set as TRUE, an extra variable "file_name" will be added to each data.frame in the output list.

}
