% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment0.R
\name{segment_0}
\alias{segment_0}
\title{segment_0}
\usage{
segment_0(speed, threshold, time_diff)
}
\arguments{
\item{speed}{the speed is a vector of speeds to be segmented.}

\item{threshold}{the threshold that is needed to segment the speeds. (the unit of threshold is an hour)}

\item{time_diff}{the difference of time for each data point.}
}
\value{
shift number, which is a vector which indicates the separated shift numbers, with 0 indicated time of taking a rest.
}
\description{
Define a funtion "segment0" that separate different shifts
}
