% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_1.R
\name{segment_1}
\alias{segment_1}
\title{segment_1}
\usage{
segment_1(data, speed, time_diff, threshold)
}
\arguments{
\item{data}{The data with speed and time_diff to be separated.}

\item{speed}{the variable name of speed to be segmented in the data.}

\item{time_diff}{the variable name of time_diff to weight the speed in the data.}

\item{threshold}{the threshold that is needed to segment the speeds. (the unit of threshold is an hour)}
}
\value{
shift number, which is a vector which indicates the separated shift numbers, with 0 indicated time of taking a rest.
}
\description{
A version of "segment0" that could be passed onto a pipe function. Note: "segment0" is a function that separates different shifts
}
