% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_data}
\alias{tidy_data}
\title{tidy_data}
\usage{
tidy_data(data, delete_duplicate = TRUE)
}
\arguments{
\item{data}{the data to be cleaned.}

\item{delete_duplicate}{where the duplicated data points are to be deleted. The default value is TRUE, which means the duplicated data points will be deleted by default.}
}
\description{
This function cleans the driver data and returns the cleaned dataset.
}
